fix: resolve all pyright standard-mode errors for clean CI typecheck
Exclude uot/algorithms and uot/solvers/pdlp_bary (internal algorithm
ports with JAX pytree typing patterns incompatible with standard mode)
from pyright. Fix all remaining errors across solvers, problems,
generators, experiments, and utils: isinstance narrowing for GridMeasure
vs BaseMeasure, h5py Group narrowing, cast for JAX/optax PyTree ops,
and miscellaneous annotation corrections.

pyright uot now reports 0 errors, 42 warnings (missing stubs for
scipy/sklearn/POT — acceptable for third-party packages). 158 tests pass.

+++ uot/solvers/sinkhorn/sinkhorn_log.py
@@ -31,7 +31,7 @@ class SinkhornTwoMarginalLogJaxSolver(BaseSolver):
         cost_original = costs[0]
         cost_scale = jnp.max(jnp.abs(cost_original))
         C = cost_original / cost_scale if normalize_cost else cost_original
-        reg = reg / cost_scale if normalize_cost else reg
+        reg = float(reg / cost_scale) if normalize_cost else reg
         mu, nu = marginals[0].as_point_cloud()[1], marginals[1].as_point_cloud()[1]
 
         P, cost, phi, psi, n_steps, err = sinkhorn_jax(
